Added simple path finding evolution sketch

+++ Nature of Code/Week 7/NOCBreeder.h
@@ -25,6 +25,6 @@
 - (void)mutate;
 
 + (double)mutationRate;
-+ (void)setMutationRate:(float)newRate;
++ (void)setMutationRate:(double)newRate;
 
 @end
